Initial commit for cassalog that provides some core functionality
SchemaManager is the main entry point for running cassalog. It currently
expects the schema to already exist. It will create the change log table if
necessary. Only changes that are not in the change log table will be applied.

A hash of the CQL is stored in the change log table. If a change has already
been applied, the hash in the database is compared to the hash of the CQL in
the source script. An exception is thrown if they differ. This is similar to
behavior provided by Liquibase.

The mini dsl allows you to specify an id, author, description, tags, and the
CQL for each change. Only the id and cql are (or will be) required.

Changes are applied in the order declared in the script. A revision number is
stored in the change log table for each change in order to keep track of the
order in which changes are applied and more importantly to ensure changes do
not get overwritten in the change_log table. Initially, I used the id as a
clustering column in the primary key, but since I am not doing anything to
enforce unique ids, it would be possible for a row to be overwritten
inadvertently. As long as cassalog is executed by a single client application
and not concurrently, changes will receive unqique, montonically increasing
revision numbers.