initial support for tags
You can now specify a set of tags when running cassalog. Only those change sets
having no tags or having tags that match all of the specified tags will be
executed. I have decided to somewhat model the behavior after Liquibase. A good
use case for this is seeding the database with test data in an automated
testing environment.