add createKeyspace function
This commit introduces support for creating a keyspace with a new function that
looks like,

// example script
createKeyspace {
  id 'create-cassalog_dev'
  name 'cassalog_dev'
  author 'admin'
  description 'create keyspace test'
  replication {
    strategy 'SimpleStrategy'
    replication_factor 2
  }
}

When a createKeyspace call is the first change set in a script and the keyspace
does not already exist, cassalog will create the schema and then create the
change log table in the same schema.

The ChangeSet class is now a base class with CqlChangeSet and CreateKeyspace
extending it. The validate method has been moved from CassalogService and into
the ChangeSet classes. A ChangeSet class should expose a string property named
cql that returns the CQL to execute.