change API for executing scripts, add support for passing vars to scripts
Up until now, scripts were specified as a java.io.Reader. This was done
primarily to make testing easier so that I could define scripts in line in
the test source code. In practice though scripts will be defined in separate
source files. The script argument to the execute method(s) is now a
java.net.URI. All of the tests have been refactored, moving the scripts to
external files. This has the downside of not being able to see the scripts
right there in the test method. It has the upside of not having to escape
CQL and overall making the scripts more readable.

Another big change in this commit is adding support for passing variables to
scripts. The variables can be of any type, and several of the tests pass the
keyspace name into the scripts as an example.