add support for recreating a keyspace
The createKeyspace function now has an attribute named recreate. When set to
true, Cassalog will first drop the keyspace (if it exists) before creating it.
This behavior only applies for the initial change set, that is when the
createKeyspace function is the first change set function encountered. This can
be useful in environments for running automated tests like CI builds in which
you want to start with a clean environment including a fresh keyspace.