Introduce Cassalog interface and update dsl apis
The Cassalog class has been renamed to CassalogImpl which implements the new
Cassalog interface. Having an interface is the cleanest way to present the API
available to clients. CassalogBuilder can be used to obtain an instance of
Cassalog.

All of the ChangeSet methods that are exposed to scripts and that took String
params have had their argument types changed from String to def in order to
support both String and GString arguments. In a Groovy only environment,
GString is automagically converted to a String, but when executing Cassalog
scripts from Java.