several changes to make sure we are in a consistent state
Previously the schema change would be applied and then the changelog table
would get updated. With this commit, we now first update the changelog table,
but the applied_at column is not updated. We then apply the schema change(s).
Then we set the applied_at column.

There is also a new changeset function in the cassalog scripts. It is named
verify and it accepts a closure that should return a boolean. The purpose of
this function is to verify whether or not a schema change has been applied.
Here is an example:

schemaChange {
  author 'jsanda'
  version '1.0'
  cql "CREATE TABLE test (x int PRIMARY KEY, y int)"
  verify { tableExists(keyspacec, 'test') }
}

When Cassalog encounters this schema change, it goes through the steps
described above. If the schema change is already in the changelog table, and
the applied_at column is null, then the verify function is called if it is
defined. If the function returns true, Cassalog interprets this to mean that
the schema change has already been applied and the applied_at column is set.
If the verify function returns false, the schema change is applied and then
the applied_at column is set.